Updated DWeb producer
Updated producer implementation to perform all required DWeb operations.
Added functions required for publishing new objects and advertising the
associated OIDs for successfully published objects.
The current implementation advertises 10 new objects with a 1 second
delay after starting the simulation. The delay is required in order for
the UDP client to work correctly. Further work is required to advertise
new objects throughout the simulation.

The producer stores the associated OID and metadata in a map for each
object it publishes. Using this, objects are reconstructed when an
interest is received. Note this behaviour is important because it differs
from the ordinary ndnSIM behaviour where objects are generated upon
receiving an interest. As DWeb objects must be published and stored in
the blockchain before a consumer can request them, the producer must
create the object in advance. The use of the map to recreate the objects
eliminates the need to store them.