Rewrote MAGIC implementation
Re-implemented MAGIC to use popularity values in the interest and data
packets.
Popularity is stored in the ApplicationParams in the interest packet.
Required modifying the interest packet to remove feature which appended a
hash of the ApplicationParams to the name, thereby corrupting the request.

The maximum popularity is added to the MetaInfo field in the returned data
packet.

Added MAGICParams class for reading and writing MAGIC parameters to and
from data and interest packets.