Update UDP client
Removed previous ns-3 based UDP client and replaced it with an
asynchronous boost UDP client.
The ns-3 based client contained a bug which prevented the socket from
working correctly when scaling up the simulation. Increasing the number of
nodes in the simulation would result in packets no longer arriving in the
docker container.
Additional updates were required in the python server to handle the new
communication process on seperate ports.