Update MAGIC for final results
This commit contains changes made over the last several weeks to tweak
performance for the final results. Updated the popularity to consider
the last 15 seconds in order to better track objects with very few
request.
Corrected bug in caching behaviour to ensure caching occurs when local
gain exceeds the max on the path.